Use uint64_t add to export to dot

+++ src/storm-dft/api/storm-dft.cpp
@@ -15,11 +15,11 @@ namespace storm {
         void exportDFTToBddDot(std::shared_ptr<storm::storage::DFT<double>> const& dft, storm::dd::DdType const type, std::string const& file) {
             if(type == storm::dd::DdType::CUDD) {
                 storm::modelchecker::SFTBDDChecker<double, storm::dd::DdType::CUDD> checker{dft};
-                checker.translate().template toAdd<double>().exportToDot(file);
+                checker.translate().template toAdd<uint64_t>().exportToDot(file);
             }
             else if(type == storm::dd::DdType::Sylvan) {
                 storm::modelchecker::SFTBDDChecker<double, storm::dd::DdType::Sylvan> checker{dft};
-                checker.translate().template toAdd<double>().exportToDot(file);
+                checker.translate().template toAdd<uint64_t>().exportToDot(file);
             } else {
                 STORM_LOG_THROW(false, storm::exceptions::NotSupportedException, "Export to Bdd Dot not supported for this dd library.");
             }
